requirements.txt
qdrant-client>=1.17.0
openai>=1.3.0
pdfplumber>=0.10.3
streamlit>=1.28.0
fastapi>=0.104.1
uvicorn>=0.24.0
python-dotenv>=1.0.0
numpy>=1.24.0
